+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file. The format 
 - Renamed the heading and admin tab for "KPIs" to "measurements".
 - The graphs trend tag does now differ on display type 'integer' and 'percentage'
   when displaying the increase/decrease for period.
+- Current goals are now listed on the KPI admin page.
 
 ### Removed